Move globalState type to types dir
